Added ssl certificate validation and authentication
Python don't validate the ssl certificate of any hosts. Now you can
provide a CA-file to validate the webservers certificate.

Furthermore you can do ssl based client certificate authentication
against NSCAweb and NRDP destinations. Sadly the CherryPy http framework
do not provide such a functionality in server mode, so you need a
webserver which can provide this and passes the requests to
NSCAweb/NRDP. (see mod_proxy for Apache)

In the config file there are 3 new optional options:
  - sslkeyfile: the Key-File (Client)
  - sslcertfile: the Certificate (Client)
  - sslcafile: the Public CA Key to validate the servers certificate

The sslkeyfile and sslcertfile can be the same file, if it proviedes
both, the key and certificate of the client. It should be not protected
by a pass parse!
The sslcafile provides the public key of the servers certificate. If it
isn't setted, the validation will be disabled.